Include Http Server
Since we are going to receive http requests from our clients, it was
necessary to create a http server to setup a socket, and listen to
clients request to this socket.

After receiveing the request, the httpServer will get the client socket,
and prepare a response. After respoding to the client, the httpserver
closes the connection and keep alive to listen and interact with other
clients.